新闻安全/requirements.txt
requests
pandas
beautifulsoup4
praw
google-api-python-client
httplib2
PySocks
python-dotenv
selenium
aiohttp
numpy
openpyxl
wordcloud
matplotlib
jieba
nltk
langdetect
snownlp
gensim
pyLDAvis
scikit-learn